Eliminate hardcode of FIPS_KERNEL_RAM_BASE in the decompressor.
The implementation of the FIPS integrity check consists of a modification
to the Linux decompressor whereby the (compressed) zImage is copied to a
special location in RAM prior to decompression.  Unfortunately this
implementation relies on a hardcoded FIPS_KERNEL_RAM_BASE address, which
breaks the position-independent property of the decompressor, and thus,
breaks kexec hardboot.

Here we eliminate FIPS_KERNEL_RAM_BASE in favor of computing the compressed
kernel base address, which is the start of the linked .text segment, using
position independent code.